Refactor to remove cyclic imports
Define `TractorFullCatalog` in des.py. So now DES can be thought of to
be entirely encapsulated by DECaLS.
